+++ ring-core/src/ring/middleware/head.clj
@@ -6,6 +6,7 @@
 
 (defn head-request
   "Turns a HEAD request into a GET."
+  {:added "1.2"}
   [request]
   (if (= :head (:request-method request))
     (assoc request :request-method :get)
@@ -13,6 +14,7 @@
 
 (defn head-response
   "Returns a nil body if original request was a HEAD."
+  {:added "1.2"}
   [response request]
   (if (and response (= :head (:request-method request)))
     (assoc response :body nil)
@@ -21,6 +23,7 @@
 (defn wrap-head
   "Middleware that turns any HEAD request into a GET, and then sets the response
   body to nil."
+  {:added "1.1"}
   [handler]
   (fn [request]
     (-> request